Refactor config.go
Move the configuration structs and routines to their own go file.
